Bugfix for kerker and initial chage density after DM in spin polarized calculations (Spin.FixSpin F)
In the case of spin-polarized calculations with Spin.FixSpin F, "Kerker" scheme does not work. (SC.KerkerPreCondition T)
Initial charge after the movement of atoms is reset, using the atomic charge with the initial setting of spin.
even after the matrix elements are correctly read from the information of the previous step. (AtomMove.ReuseDM T)
The problem of 1) comes from that the change of "G=0" term are not considered in the present code.
(G=0 term : sum of electron number, for up and down spin.)
For 2), ne_spin_in_cell is not updated during the SCF cycle (even after we reach SCF),
and spin polarization from the initial charge density made by the last-step DM does
not agree with ne_spin_in_cell.
Then, initial charge density is made from atomic charge density with the initially determined ne_spin_in_cell.